+++ front/app/modules/commercial/customizable_navbar/admin/containers/EditPageForm/EditPageFormNotInNavbar.tsx
@@ -1,0 +1,118 @@
+import React from 'react';
+import styled from 'styled-components';
+import { withRouter, WithRouterProps } from 'react-router';
+import { Formik, FormikProps } from 'formik';
+
+// components
+import PageForm, { validatePageForm, FormValues } from 'components/PageForm';
+import GoBackButton from 'components/UI/GoBackButton';
+import T from 'components/T';
+
+// utils
+import { isNilOrError } from 'utils/helperUtils';
+import { fontSizes } from 'utils/styleUtils';
+import clHistory from 'utils/cl-router/history';
+import { NAVIGATION_PATH } from '..';
+
+// services
+import { updatePage } from '../../../services/pages';
+import { handleAddPageFiles, handleRemovePageFiles } from 'services/pageFiles';
+
+// hooks
+import useAppConfigurationLocales from 'hooks/useAppConfigurationLocales';
+import useRemoteFiles from 'hooks/useRemoteFiles';
+import usePage from 'hooks/usePage';
+import usePageSlugs from 'hooks/usePageSlugs';
+
+const Title = styled.h1`
+  font-size: ${fontSizes.xxxl}px;
+  padding: 0;
+  width: 100%;
+  margin: 1rem 0 3rem 0;
+`;
+
+const EditPageFormNotInNavbar = ({ params: { pageId } }: WithRouterProps) => {
+  const appConfigurationLocales = useAppConfigurationLocales();
+  const page = usePage({ pageId });
+  const remotePageFiles = useRemoteFiles({
+    resourceType: 'page',
+    resourceId: !isNilOrError(page) ? page.id : null,
+  });
+  const pageSlugs = usePageSlugs();
+
+  if (
+    isNilOrError(page) ||
+    isNilOrError(appConfigurationLocales) ||
+    isNilOrError(pageSlugs)
+  ) {
+    return null;
+  }
+
+  const handleSubmit = async (
+    { local_page_files, ...pageUpdate }: FormValues,
+    { setSubmitting, setStatus }
+  ) => {
+    try {
+      const promises: Promise<any>[] = [updatePage(pageId, pageUpdate)];
+
+      if (!isNilOrError(local_page_files)) {
+        const addPromise = handleAddPageFiles(
+          pageId,
+          local_page_files,
+          remotePageFiles
+        );
+        const removePromise = handleRemovePageFiles(
+          pageId,
+          local_page_files,
+          remotePageFiles
+        );
+
+        promises.push(addPromise, removePromise);
+      }
+
+      await Promise.all(promises);
+
+      setStatus('success');
+      setSubmitting(false);
+    } catch (error) {
+      setStatus('error');
+      setSubmitting(false);
+    }
+  };
+
+  const goBack = () => {
+    clHistory.push(NAVIGATION_PATH);
+  };
+
+  const renderFn = (props: FormikProps<FormValues>) => {
+    return <PageForm {...props} pageId={pageId} />;
+  };
+
+  return (
+    <div>
+      <GoBackButton onClick={goBack} />
+      <Title>
+        <T value={page.attributes.title_multiloc} />
+      </Title>
+      <Formik
+        initialValues={{
+          title_multiloc: page.attributes.title_multiloc,
+          body_multiloc: page.attributes.body_multiloc,
+          slug: page.attributes.slug,
+          local_page_files: remotePageFiles,
+        }}
+        onSubmit={handleSubmit}
+        render={renderFn}
+        validate={validatePageForm(
+          appConfigurationLocales,
+          pageSlugs,
+          page.attributes.slug
+        )}
+        validateOnChange={false}
+        validateOnBlur={false}
+      />
+    </div>
+  );
+};
+
+export default withRouter(EditPageFormNotInNavbar);