+++ front/app/modules/commercial/customizable_navbar/admin/containers/VisibleNavbarItemList/VisibleNavbarItemList.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render, screen, fireEvent } from 'utils/testUtils/rtl';
-import VisibleNavbarItemList from './VisibleNavbarItemList';
-import { reorderNavbarItem, removeNavbarItem } from '../../services/navbar';
-import { deletePage } from 'services/pages';
+import VisibleNavbarItemList from '.';
+import { reorderNavbarItem, removeNavbarItem } from '../../../services/navbar';
+import { deletePage } from '../../../services/pages';
 
 jest.mock('services/locale');
 jest.mock('services/appConfiguration');
@@ -10,12 +10,12 @@ jest.mock('hooks/useNavbarItems');
 jest.mock('hooks/usePageSlugById');
 jest.mock('hooks/useLocale');
 
-jest.mock('../../services/navbar', () => ({
+jest.mock('../../../services/navbar', () => ({
   reorderNavbarItem: jest.fn(),
   removeNavbarItem: jest.fn(),
 }));
 
-jest.mock('services/pages', () => ({
+jest.mock('../../../services/pages', () => ({
   deletePage: jest.fn(),
 }));