+++ front/app/modules/commercial/customizable_navbar/admin/containers/VisibleNavbarItemList/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 // services
-import { reorderNavbarItem, removeNavbarItem } from '../../services/navbar';
-import { deletePage } from 'services/pages';
+import { reorderNavbarItem, removeNavbarItem } from '../../../services/navbar';
+import { deletePage } from '../../../services/pages';
 import { INavbarItem, getNavbarItemSlug } from 'services/navbar';
 
 // components
@@ -12,7 +12,7 @@ import {
   LockedRow,
 } from 'components/admin/ResourceList';
 import { SubSectionTitle } from 'components/admin/Section';
-import NavbarItemRow from '../components/NavbarItemRow';
+import NavbarItemRow from '../../components/NavbarItemRow';
 
 // hooks
 import useNavbarItems from 'hooks/useNavbarItems';
@@ -22,10 +22,12 @@ import usePageSlugById from 'hooks/usePageSlugById';
 // i18n
 import { injectIntl, FormattedMessage } from 'utils/cl-intl';
 import { InjectedIntlProps } from 'react-intl';
-import messages from './messages';
+import messages from '../messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
+import clHistory from 'utils/cl-router/history';
+import { NAVIGATION_PATH } from '../';
 
 const VisibleNavbarItemList = ({
   intl: { formatMessage },
@@ -41,6 +43,14 @@ const VisibleNavbarItemList = ({
   ) {
     return null;
   }
+
+  const handleClickEdit = (navbarItem: INavbarItem) => () => {
+    const pageData = navbarItem.relationships.static_page.data;
+
+    pageData
+      ? clHistory.push(`${NAVIGATION_PATH}/pages/edit/${pageData.id}`)
+      : clHistory.push(`${NAVIGATION_PATH}/navbar-items/edit/${navbarItem.id}`);
+  };
 
   const handleClickView = (navbarItem: INavbarItem) => () => {
     const originWithLocale = `${window.location.origin}/${locale}`;
@@ -88,6 +98,8 @@ const VisibleNavbarItemList = ({
                 <NavbarItemRow
                   title={navbarItem.attributes.title_multiloc}
                   isDefaultPage={navbarItem.attributes.code !== 'custom'}
+                  showEditButton={navbarItem.attributes.code !== 'home'}
+                  onClickEditButton={handleClickEdit(navbarItem)}
                   onClickViewButton={handleClickView(navbarItem)}
                 />
               </LockedRow>
@@ -105,7 +117,9 @@ const VisibleNavbarItemList = ({
                 <NavbarItemRow
                   title={navbarItem.attributes.title_multiloc}
                   isDefaultPage={navbarItem.attributes.code !== 'custom'}
+                  showEditButton
                   showRemoveButton
+                  onClickEditButton={handleClickEdit(navbarItem)}
                   onClickRemoveButton={handleClickRemove(navbarItem.id)}
                   onClickDeleteButton={handleClickDelete(
                     navbarItem.relationships.static_page.data?.id